Add JSON structure description
Also, to respond to my own comment in the last commit,
tree-sitter can also output a visual graph for the node tree
in SVG format, which is way more readable than unformatted sexp.
Still, it omits necessary details and the graph wastes a lot of visual
space so the image easily gets really large for more slightly
more complex tree. JSON is a nice compact tree visualization
format, if it's structured properly.